requirement.txt
mysqlclient
flask
flask_sqlalchemy
openai
flask_restx
flask_cors
pymysql
pyjwt